refactor(utils): add util functions in isolated folders
