+++ src/shared/utils/hash/index.ts
