+++ src/shared/utils/index.ts
@@ -1,4 +1,4 @@
 export * from './dayjs';
 export * from './generate-data';
-export * from './generateUUID';
+export * from './generate-uuid';
 export * from './hash';
